requirements.txt
dlib
imageio==2.34.1
joblib==1.4.2
lazy_loader==0.4
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pygame==2.5.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tzdata==2024.1